requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu # cpu version to reduce the slug size on heroku
huggingface-hub==0.20.3
joblib==1.2.0
numpy==1.24.3
pandas==2.1.4
python-telegram-bot==21.1.1
scikit-learn==1.3.0
torcheval==0.0.7
tqdm==4.65.0
transformers==4.40.2
python-dotenv==1.0.1
gradio_client==0.16.4